+++ src/oasapi/__init__.py
@@ -1,6 +1,6 @@
 __version__ = "0.1.11"
 
-from .transform import prune_unused_global_items, prune_unused_security_definitions
-from .validation import validate_swagger
+from .prune import prune_unused_global_items, prune_unused_security_definitions
+from .validation import validate
 
-__all__ = ["validate_swagger", "prune_unused_global_items", "prune_unused_security_definitions"]
+__all__ = ["validate", "prune_unused_global_items", "prune_unused_security_definitions"]